% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_calculated.R
\name{as_calculated}
\alias{as.calculated}
\alias{as_calculated}
\alias{is.calculated}
\alias{is_calculated}
\title{Conver to Calculated Object}
\usage{
as_calculated(x, sigfig, label = character(0), units = character(0))

as.calculated(x, sigfig, label = character(0), units = character(0))

is_calculated(x)

is.calculated(x)
}
\arguments{
\item{x}{For \code{as_calculated} A vector of either class \code{numeric} 
or \code{double}. Vectors of other classes may be given, but they will 
be returned unchanged.  For \code{is_calculated} an object of to be tested.}

\item{sigfig}{A vector of integerish values. This vector will be recycled 
over the length of \code{x}.  The value specifies the number of 
significant figures for the calculated value.}

\item{label}{\code{character(1)}. The human-friendly label for the vector.}

\item{units}{\code{character(1)}. Gives the unit of measurement.}
}
\description{
Values that result from calculations involving measurements
  are subject to imprecision based on the imprecision of the values 
  used in the calculation.  The \code{calculated} class tracks the 
  the abstract precision and the significant figures associated with the 
  calculated values, allowing them to be used in subsequent calculations 
  or for printing.
}
\section{Functional Requirements}{

\enumerate{
  \item  If \code{x} is \code{numeric} or \code{double}, return a vector
    with \code{sigfig}, \code{label}, and \code{units} attributes.
  \item If \code{sigfig} is not integerish, cast an error.
  \item \code{sigfig} may have length no greater than \code{x}.
  \item If \code{length(x)} is not a multiple of \code{length(sigfig)}, cast a 
    warning.
  \item If \code{label} is not a character vector with length no greater 
    than 1, cast an error.
  \item If \code{units} is not a character vector with length no greater 
    than 1, cast an error.
  \item If \code{x} is an integer, cast a warning that no action is 
    taken because integers have no measurement imprecision.
  \item If \code{x} is not \code{numeric}, \code{double}, or \code{integer},
    cast a warning that no action is taken.
  \item If \code{x} is not \code{numeric} or \code{double},
    the vector returned is identical to \code{x}
  \item If any \code{sigfig} is less than 1, cast an error.
}
}
\author{
Benjamin Nutter
}

