% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_measured.R
\name{extract_measured}
\alias{[.calculated}
\alias{[.measured}
\alias{[<-.calculated}
\alias{[<-.measured}
\alias{extract_measured}
\title{Extraction Methods for Measured and Calculated Values}
\usage{
\method{[}{measured}(x, i)

\method{[}{measured}(x, i) <- value

\method{[}{calculated}(x, i)

\method{[}{calculated}(x, i) <- value
}
\arguments{
\item{x}{A \code{measured} or \code{calculated} object.}

\item{i}{A vector of indices to extract or replace}

\item{value}{replacement values for \code{x}. If \code{values} has 
\code{precision} or \code{sigfig} attributes, they will replace the 
attributes at positions \code{i} in \code{x}. Otherwise, only the 
attributes are left as they are.}
}
\description{
Extraction and replacement methods for measured and
  calculated objects.
}
\section{Functional Requirements}{

Extraction Method \cr
\enumerate{
  \item The values at positions \code{i} are extracted and retain 
    their attributes.
}
Replacement Method \cr
\enumerate{
  \item When \code{value} has a \code{precision} attribute, the 
      \code{precision} attribute of \code{x} at positions \code{i}
      are replaced by the \code{precision} attribute of \code{value}
  \item When \code{value} has no \code{precision} attribute, the 
      value of \code{x} is replaced with no change to the 
      \code{precision} attribute of \code{x}.
  \item When \code{value} has a \code{sigfig} attribute, the 
      \code{sigfig} attribute of \code{x} at positions \code{i}
      are replaced by the \code{sigfig} attribute of \code{value}
  \item When \code{value} has no \code{sigfig} attribute, the 
      value of \code{x} is replaced with no change to the 
      \code{sigfig} attribute of \code{x}.
  \item Replacement works with named vectors.
}
}
\author{
Benjamin Nutter
}

