% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.calculated.R
\name{print.calculated}
\alias{print.calculated}
\title{Print Calculated Vector}
\usage{
\method{print}{calculated}(x, label = getOption("imprecise_label", FALSE),
  units = getOption("imprecise_units", FALSE), ...)
}
\arguments{
\item{x}{A vector that inherits class \code{calculated}}

\item{label}{\code{logical(1)}. When \code{TRUE}, the label is printed.}

\item{units}{\code{logical(1)}. When \code{TRUE}, the unit of measurement
is printed.}

\item{...}{Arguments to pass to other methods.}
}
\description{
Print a Calculated Vector
}
\section{Functional Requirements}{

\enumerate{
  \item The value of \code{x} is printed with the significant figures 
    in the \code{sigfig} attribute.
  \item When \code{label = TRUE}, the \code{label} attribute is 
    printed above the output.
  \item When \code{units = TRUE}, the \code{units} attribute is 
    printed above the output.
  \item when \code{units = TRUE} and there are no non-white space 
    characters in the units attribute, quietly reset \code{units}
    to \code{FALSE}
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{print.measured}}
}

