% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.label.R
\name{print.label}
\alias{print.label}
\title{Print a Vector with a Label Attribute}
\usage{
\method{print}{label}(x, label = getOption("imprecise_label", FALSE),
  units = getOption("imprecise_units", FALSE), ...)
}
\arguments{
\item{x}{An atomic vector to be printed.}

\item{label}{\code{logical(1)}. When \code{TRUE}, the label is printed 
above the vector.}

\item{units}{\code{logical(1)}. When \code{TRUE}, the units are printed
above the vector, appended to the label if it is printed.}

\item{...}{Additional arguments to pass to other methods.}
}
\description{
Vectors with a \code{label} subclass is, by default, 
  printed as if it did not have the \code{label} subclass.  The label 
  may be printed by using \code{options(imprecise_label = TRUE)}.
}
\section{Functional Requirements}{

\enumerate{
  \item prints the vector using the method that would be used if the 
    vector did not carry the \code{label} subclass.
  \item When \code{label = TRUE}, the \code{label} attribute is 
    printed above the output.
  \item When \code{units = TRUE}, the \code{units} attribute is 
    printed above the output.
  \item when \code{units = TRUE} and there are no non-white space 
    characters in the units attribute, quietly reset \code{units}
    to \code{FALSE}
}
}
\author{
Benjamin Nutter
}

