% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.measured.R
\name{print.measured}
\alias{print.measured}
\title{Print Measured Vector}
\usage{
\method{print}{measured}(x, label = getOption("measured_label", FALSE),
  units = getOption("measured_units", FALSE), ...)
}
\arguments{
\item{x}{A vector that inherits class \code{measured}}

\item{label}{\code{logical(1)}. When \code{TRUE}, the label is printed.}

\item{units}{\code{logical(1)}. When \code{TRUE}, the unit of measurement
is printed.}

\item{...}{Arguments to pass to other methods.}
}
\description{
Print a Measured Vector
}
\section{Functional Requirements}{

\enumerate{
  \item The value of \code{x} is rounded to the corresponding value 
    in the \code{precision} attribute.
  \item When \code{label = TRUE}, the \code{label} attribute is 
    printed above the output.
  \item When \code{units = TRUE}, the \code{units} attribute is 
    printed above the output.
  \item when \code{units = TRUE} and there are no non-white space 
    characters in the units attribute, quietly reset \code{units}
    to \code{FALSE}
}
}
\author{
Benjamin Nutter
}

