% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_label.R
\name{set_label}
\alias{get_label}
\alias{get_units}
\alias{set_label}
\alias{set_units}
\title{Set and Get Label and Units Attributes}
\usage{
set_label(x, label)

get_label(x)

set_units(x, units)

get_units(x)
}
\arguments{
\item{x}{An atomic vector.}

\item{label}{\code{character(1)} for the \code{label} attribute.}

\item{units}{\code{character(1)} for the \code{units} attribute.}
}
\description{
Set and retrieve labels and units from vectors.  While
  the \code{measured} and \code{calculated} classes explicitly 
  define an attribute for the label and units, they may be added to 
  any atomic vector.  These behave similarly to
}
\section{Functional Requirements}{

\enumerate{
  \item If \code{x} is not an atomic vector, cast an error.
  \item For the \code{set} methods, assign a new attribute, 
    \code{label} or \code{units} and give it 
    the value of the corresponding argument.
  \item For \code{get_label}, if the \code{label} attribute is 
    \code{NULL} or doesn't exist, return the name of \code{x}
  \item For \code{get_units}, if the \code{units} attribute is 
    \code{NULL} or doesn't exist, return an empty string. 
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{set_precision}}, \code{\link{get_precision}},
  \code{\link{set_sigfig}}, \code{\link{get_sigfig}}
}

