% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_precision.R
\name{set_precision}
\alias{get_precision}
\alias{set_precision}
\title{Set and Get Precision Attribute of Measured Objects}
\usage{
set_precision(x, precision)

get_precision(x)
}
\arguments{
\item{x}{An object of class \code{measured}.}

\item{precision}{A vector of integerish values}
}
\description{
Set or retrieve the precision attribute of a measured object.
}
\section{Functional Requirements}{

\code{set_precision}\cr
\enumerate{
  \item If \code{x} is not a \code{measured} object, cast a warning and 
    return the vector unchanged.
  \item Correctly resets the precision attribute.
  \item If \code{length(x)} is not a multiple of \code{length(precision)},
    cast a warning.
  \item If \code{length(precision)} is greater than \code{length(x)},
    cast an error.
}

\code{get_precision}\cr
\enumerate{
  \item Retrieves the correct precision values for \code{measured} 
      and \code{calculated} objects.
  \item If \code{x} is not \code{measured} or \code{calculated}, 
      return \code{NULL} and cast a warning.
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{set_sigfig}}, \code{\link{get_sigfig}}, 
  \code{\link{set_label}}, \code{\link{get_label}},
  \code{\link{set_units}}, \code{\link{get_units}}
}

