% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sigfig.R
\name{set_sigfig}
\alias{get_sigfig}
\alias{set_sigfig}
\title{Set and Get Significant Figure Attribute of Measured Objects}
\usage{
set_sigfig(x, sigfig)

get_sigfig(x)
}
\arguments{
\item{x}{An object of class \code{calculated}.}

\item{sigfig}{A vector of integerish values.}
}
\description{
Set or retrieve the sigfig attribute of a measured object.
}
\section{Functional Requirements}{

\code{set_sigfig}\cr
\enumerate{
  \item If \code{x} is not a \code{calculated} object, cast a warning and 
    return the vector unchanged.
  \item Correctly resets the sigfig attribute.
  \item If \code{length(x)} is not a multiple of \code{length(sigfig)},
    cast a warning.
  \item If \code{length(sigfig)} is greater than \code{length(x)},
    cast an error.
  \item If and \code{sigfig} is less than 1, cast an error.
}

\code{get_sigfig}\cr
\enumerate{
  \item Retrieves the correct precision values for a \code{calculated} object.
  \item If \code{x} is not \code{calculated}, return \code{NULL} and 
    cast a warning.
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{set_precision}}, \code{\link{get_precision}}, 
  \code{\link{set_label}}, \code{\link{get_label}},
  \code{\link{set_units}}, \code{\link{get_units}}
}

