% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.measured.R
\name{format.measured}
\alias{format.calculated}
\alias{format.measured}
\title{Format Measured and Calculated Objects}
\usage{
\method{format}{measured}(x, ...)

\method{format}{calculated}(x, ...)
}
\arguments{
\item{x}{An object of class \code{measured} or \code{calculated}}

\item{...}{Additional arguments to pass to other methods.}
}
\description{
Print \code{measured} and \code{calculated} objects to 
  either the appropriate precision or significant figures. The primary
  motivation for these functions is to print measured and calculated 
  values appropriately in data frames.
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{format}}
}

